+++ assets/src/scripts/schema.js
@@ -57,7 +57,7 @@ const timeSeries = memoize(tsKey => new schema.Entity('timeSeries', {
                 const value = parseFloat(v.value);
                 return {
                     ...v,
-                    dateTime: new Date(v.dateTime),
+                    dateTime: new Date(v.dateTime).getTime(),
                     value: value === parent.variable.noDataValue ? null : value
                 };
             })
@@ -99,8 +99,8 @@ const queryInfo = memoize(tsKey => new schema.Entity('queryInfo', {}, {
                 mode: parts[1],
                 modifiedSince: parts[2] === 'null' ? null : parts[3],
                 interval: {
-                    start: new Date(parts[3]),
-                    end: new Date(parts[4])
+                    start: new Date(parts[3]).getTime(),
+                    end: new Date(parts[4]).getTime()
                 }
             };
         } else {
@@ -108,7 +108,7 @@ const queryInfo = memoize(tsKey => new schema.Entity('queryInfo', {}, {
         }
 
         // Store requestDT as a date object
-        queryInfo.notes.requestDT = new Date(queryInfo.notes.requestDT);
+        queryInfo.notes.requestDT = new Date(queryInfo.notes.requestDT).getTime();
 
         return queryInfo;
     }