+++ assets/src/scripts/selectors/timeSeriesSelector.spec.js
@@ -1,6 +1,6 @@
 const { getVariables, getCurrentVariableID, getCurrentDateRange, getCurrentVariable, getQueryInfo, getRequests,
     getCurrentParmCd, hasTimeSeries, getTsRequestKey, getTsQueryInfo, getRequestTimeRange, isLoadingTS, getTSRequest,
-    getTimeSeriesCollectionIds } = require('./timeSeriesSelector');
+    getTimeSeriesCollectionIds, getIanaTimeZone, getNwisTimeZone} = require('./timeSeriesSelector');
 
 describe('timeSeriesSelector', () => {
     const TEST_VARS = {
@@ -259,7 +259,7 @@ describe('timeSeriesSelector', () => {
                 queryInfo: {
                     'current:P7D': {
                         notes: {
-                            requestDT: new Date('2017-03-31'),
+                            requestDT: 1490936400000,
                             'filter:timeRange': {
                                 mode: 'PERIOD',
                                 periodDays: 7,
@@ -269,12 +269,12 @@ describe('timeSeriesSelector', () => {
                     },
                     'current:P30D:00060': {
                         notes: {
-                            requestDT: new Date('2017-03-31'),
+                            requestDT: 1490936400000,
                             'filter:timeRange': {
                                 mode: 'RANGE',
                                 interval: {
-                                    start: new Date('2017-03-01'),
-                                    end: new Date('2017-03-29')
+                                    start: 1488348000000,
+                                    end: 1490763600000
                                 }
                             }
                         }
@@ -307,7 +307,7 @@ describe('timeSeriesSelector', () => {
                 queryInfo: {
                     'current:P7D': {
                         notes: {
-                            requestDT: new Date('2017-03-31'),
+                            requestDT: 1490936400000,
                             'filter:timeRange': {
                                 mode: 'PERIOD',
                                 periodDays: 7,
@@ -317,12 +317,12 @@ describe('timeSeriesSelector', () => {
                     },
                     'current:P30D:00060': {
                         notes: {
-                            requestDT: new Date('2017-03-31'),
+                            requestDT: 1490936400000,
                             'filter:timeRange': {
                                 mode: 'RANGE',
                                 interval: {
-                                    start: new Date('2017-03-01'),
-                                    end: new Date('2017-03-29')
+                                    start: 1488348000000,
+                                    end: 1490763600000
                                 }
                             }
                         }
@@ -351,15 +351,63 @@ describe('timeSeriesSelector', () => {
 
         it('should use the requestDT for requests with mode PERIOD', () => {
             expect(getRequestTimeRange('current')(TEST_DATA)).toEqual({
-                start: new Date('2017-03-24'),
-                end: new Date('2017-03-31')
+                start: 1490331600000,
+                end: 1490936400000
             });
         });
 
         it('should use the interval for request with mode RANGE', () => {
             expect(getRequestTimeRange('current', 'P30D', '00060')(TEST_DATA)).toEqual({
-                start: new Date('2017-03-01'),
-                end: new Date('2017-03-29')
+                start: 1488348000000,
+                end: 1490763600000
+            });
+        });
+    });
+
+    describe('getIanaTimeZone', () => {
+
+        it('returns null if series is empty', () => {
+            expect(getIanaTimeZone({
+                series: {}
+            })).toBeNull();
+        });
+
+        it('returns the time zone when present', () => {
+            expect(getIanaTimeZone({
+                series: {
+                    ianaTimeZone: 'America/Los_Angeles'
+                }
+            })).toEqual('America/Los_Angeles');
+        });
+    });
+
+    describe('getNwisTimeZone', () => {
+
+        it('returns an empty object if series is empty', () => {
+            expect(getNwisTimeZone({
+                series: {}
+            })).toEqual({});
+        });
+
+        it('returns the NWIS provide timezones when present', () => {
+            expect(getNwisTimeZone({
+                series: {
+                    timeZones: {
+                        CDT: {
+                            content: 'x'
+                        },
+                        CST: {
+                            content: 'y'
+                        }
+                    }
+                }
+            })).toEqual({
+                CDT: {
+                    content: 'x'
+                },
+                CST: {
+                    content: 'y'
+                }
             });
         });
     });